feat: move codes into src folder
