+++ manifest.json
@@ -1,11 +1,11 @@
 {
-	"id": "ryo-usememos-sync",
+	"id": "obsidian-memos-sync",
 	"name": "Memos Sync",
 	"version": "0.3.0",
 	"minAppVersion": "0.15.0",
 	"description": "Syncing Memos to Obsidian daily note. Fully compatible with official Daily Notes plugin, Calendar plugin and Periodic Notes plugin.",
-	"author": "Obsidian",
-	"authorUrl": "https://obsidian.md",
-	"fundingUrl": "https://obsidian.md/pricing",
-	"isDesktopOnly": false
+	"author": "RyoJerryYu",
+	"authorUrl": "https://github.com/RyoJerryYu",
+	"isDesktopOnly": true,
+	"js": "main.js"
 }
