+++ src/api/memos-proto-v0.22.0/api/v1/resource_service.proto
