+++ src/api/memos-proto-v0.22.0/api/v1/workspace_service.proto
