+++ src/api/memos-proto-v0.22.0/gen/google/api/launch_stage.ts
