+++ src/api/memos-proto-v0.22.0/gen/google/protobuf/duration.ts
