+++ src/types/PluginSettings.ts
